refactor(WTEL-9812): move camelizeKeys to utils, type RawThreadPreview from ThreadPreview
- relocate generic camelizeKeys helper from task.ts to socket/utils.ts
- RawThreadPreview field value types now derived from ThreadPreview
  (snake_case keys, upstream-tied value types) instead of unknown

Refs: WTEL-9812
